requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5

# Database
azure-cosmos==4.9.0

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Email
sendgrid==6.11.0

# Scheduling
apscheduler==3.10.4

# Markdown Rendering
mistune==3.1.1

# HTMX / Static
aiofiles==24.1.0

# Security
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Linting
ruff==0.8.6
